joule_parser

Struct Parser

pub struct Parser { /* private fields */ }

Parser for Joule source code

Implementations

impl Parser

pub fn parse_attributes(&mut self) -> Result<Vec<Attribute>, ParseError>

Parse a sequence of attributes: #[attr1] #[attr2] ...

pub fn parse_attribute_args(&mut self) -> Result<Vec<AttributeArg>, ParseError>

Parse attribute arguments: key = value, key2 = value2, ...

pub fn extract_energy_budget(attributes: &[Attribute]) -> Option<EnergyBudget>

Extract EnergyBudget from a list of attributes

This method looks for #[energy_budget(...)] attributes and extracts the energy budget constraints.

pub fn extract_proc_macro_kind( attributes: &[Attribute], ) -> Option<ProcMacroKind>

Extract proc macro kind from a list of attributes.

Looks for #[proc_macro], #[proc_macro_derive(...)], or #[proc_macro_attribute].

impl Parser

pub fn new(tokens: Vec<SpannedToken>) -> Self

Create a new parser from a list of tokens

pub fn parse_file(self) -> Result<File, Vec<Diagnostic>>

Parse a complete source file

pub fn parse_single_expr(self) -> Result<Expr, Vec<Diagnostic>>

Parse a single expression (for proc macro re-parsing)

pub fn diagnostics(&self) -> &[Diagnostic]

Get accumulated diagnostics

pub fn parse_function(&mut self) -> Result<Function, ParseError>

pub fn parse_function_with_attributes( &mut self, attributes: Vec<Attribute>, ) -> Result<Function, ParseError>

pub fn parse_struct(&mut self) -> Result<Struct, ParseError>

pub fn parse_enum(&mut self) -> Result<Enum, ParseError>

pub fn parse_trait(&mut self) -> Result<Trait, ParseError>

pub fn parse_const(&mut self) -> Result<Const, ParseError>

pub fn parse_static(&mut self) -> Result<Static, ParseError>